fix: fix footer layout
